setup: separate setting up development tools from installing basic tools
Only a few packages are required to be installed on the hypervisor
machine in order to run the POC. This patch separates the old
'install_user' subroutine into 'install_user' and 'install_devel'.

'install_devel' is just what 'install_user' previously was. The new
'install_user' subroutine now only installs the most necessary packages
needed for just quickly running the POC by a new user (as quickly as
possible).

Note, that this is only a temporary change. It will most likely be
changed in future patches.