target-vm: remove the 'start local' option
This POC is currently being used primarily as a test environment for
NVMe/TCP boot. For this use case, only one extra NVMe drive is required,
and it is the one served to the client for the purpose of booting.

A future patch shall add an option to add more NVMe drives to the target
VM at startup.

Signed-off-by: Michal Rábek <mrabek@redhat.com>