setup: move legacy and developer-only functions to setup_extra.sh
Many of the steps being done in setup.sh are redundant nowdays and will
not be needed by most users of this POC project.

More refactoring is necessary. Right now, the comments and documentation
do not match the real state and behaviour of the scripts. Changes are
incoming in future patches.

Signed-off-by: Michal Rábek <mrabek@redhat.com>