target-vm: add a kickstart file for an unattended OS installation
To run the unattended installation of the target-vm, just run
'make install'. Kickstart is going to take care of all the setup that
was previously being done manually until the point when './netsetup.sh'
had to be run.

libvirt had to be used instead of QEMU in order to inject the
'anaconda-ks.cfg' file. Thus, the 'start.sh' script has also become
redundant and the VM can now be run with 'virsh start target-vm'.

'make clean' can be run to do a cleanup of everything created/modified
by the setup. That includes the VM itself registered in libvirt.