target-vm: no network connection on RHEL 10
The RHEL installation has all network interfaces disabled by default and
autoconnect is set to FALSE. For the primary interface to connect to the
network and obtain an IP address, the user is required to run
'nmcli dev connect' after every boot.

This patch solves this by enabling autoconnect immediately after the
installation via the Kickstart configuration, thus eliminating this
difference between RHEL 10 and Fedora 42.