target-vm: no need for virt-install anymore
A tool called 'livemedia-creator' from the Lorax suite was found.
It enables to inject an Anaconda Kickstart configuration file into an
ISO file. It can, however be also used to generate a qcow2 disk image
for QEMU with the ISO in the operating system already installed using
the Kickstart configuration.

Automatic partitioning could no longer be used since the utility can
only determine the required size of the disk image from the Kickstart
configuration.

The 'dracut -f --add-drivers "nvme"' was added to add the missing nvme
driver to initramfs. The reason for it missing is that the utility does
not emulate an NVMe drive while installing and thus the driver is not
added to initramfs (the system does not deem it necessary during its
generation).

Signed-off-by: Michal Rábek <mrabek@redhat.com>